+++ bin/pho.php
@@ -8,8 +8,10 @@ use Pho\Cli;
 
 $version = file_get_contents(dirname(__DIR__) . DIRECTORY_SEPARATOR . "VERSION");
 $application = new Application('Pho', $version);
+$application->add(new Cli\CompileCommand());
 $application->add(new Cli\BuildCommand());
 $application->add(new Cli\InitCommand());
+
 //$application->add(new Cli\ServeCommand());
 //$application->add(new Cli\RunCommand());
 $application->run();