+++ CHANGELOG.md
@@ -7,3 +7,8 @@ Init: will generate a folder structure
 
 ## 0.2 to 0.3
 Can now init custom projects
+
+## 0.3 to 0.4
+Brought back the old Build under the new name Compile
+TODO: let BuildCommand inherit CompileCommand
+TODO: put CompileCommand under a new family of commands: "dev"